Week 5/Exercise 14 Online Bookstore - Integration Testing for REST Services/src/main/java/com/example/bookstoreapi/test/BookControllerTest.java
package com.example.bookstoreapi.test;

import java.awt.PageAttributes.MediaType;

import org.mockito.MockitoAnnotations;
import org.springframework.beans.factory.annotation.Autowired;

import com.example.bookstoreapi.dto.BookDTO;

@WebMvcTest(BookController.class)
public class BookControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private BookService bookService;

    @BeforeEach
    public void setup() {
        MockitoAnnotations.openMocks(this);
        this.mockMvc = MockMvcBuilders.standaloneSetup(new BookController(bookService)).build();
    }

    @Test
    public void testGetBooks() throws Exception {
        BookDTO book1 = new BookDTO(1L, "Title1", "Author1", 19.99, "ISBN123");
        BookDTO book2 = new BookDTO(2L, "Title2", "Author2", 29.99, "ISBN456");

        when(bookService.getAllBooks()).thenReturn(Arrays.asList(book1, book2));

        mockMvc.perform(get("/api/books")
                .accept(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$[0].title").value("Title1"))
                .andExpect(jsonPath("$[1].title").value("Title2"));
    }
}
